Precompute some helpful information about sugyot and perek endings. This may be useful for future computations.

+++ cache_all_api_requests.ts
@@ -1,32 +1,42 @@
 /* eslint-disable no-console */
-import {ApiRequestHandler, ApiException, RealRequestMaker} from "./api_request_handler";
+import * as fs from "fs";
+import {ApiRequestHandler, RealRequestMaker} from "./api_request_handler";
 import {ApiResponse} from "./apiTypes";
 import {NoopLogger} from "./logger";
 import {books} from "./books";
 import {writeJson} from "./util/json_files";
 import {PromiseChain} from "./js/promises";
+import {cachedOutputFilePath} from "./cached_outputs";
 
 const requestHandler = new ApiRequestHandler(new RealRequestMaker());
 
 function makeRequest(bookName: string, section: string): Promise<ApiResponse> {
   return requestHandler.handleRequest(bookName, section, new NoopLogger())
     .catch(e => {
-      if (e instanceof ApiException) {
-        console.error(`ApiException in ${bookName} ${section}`);
-      }
+      console.error(`Error processing ${bookName} ${section}`);
       console.error(e);
       return makeRequest(bookName, section);
     });
 }
 
-const promiseChain = new PromiseChain();
+const promiseChains: PromiseChain[] = [];
+for (let i = 0; i < 30; i++) {
+  promiseChains[i] = new PromiseChain();
+}
 
+let i = 0;
 for (const book of Array.from(new Set(Object.values(books.byCanonicalName)))) {
   for (const section of Array.from(book.sections)) {
-    const filePath = `cached_outputs/api_request_handler/${book.canonicalName}.${section}.json`;
+    const filePath = cachedOutputFilePath(book, section);
+    if (fs.existsSync(filePath)) continue;
+    const promiseChain = promiseChains[i % promiseChains.length];
     promiseChain.add(
       () => makeRequest(book.canonicalName, section).then(result => writeJson(filePath, result)));
     promiseChain.add(() => console.log(`Finished ${book.canonicalName} ${section}`));
+    i++;
   }
 }
-promiseChain.add(() => console.log("Finished!"));
+
+for (let j = 0; j < promiseChains.length; i++) {
+  promiseChains[j].add(() => console.log(`Finished chain #${j}!`));
+}
